refactor(web): move chat controller to component
chore(web): remove head from chat controller
